refector(calculator): 클래스 분리
