+++ src/main/java/calculator/StringCalculator.java
@@ -1,11 +1,6 @@
 package calculator;
 
-import java.util.Map;
-import java.util.Optional;
-import java.util.function.IntBinaryOperator;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class StringCalculator {
 
@@ -14,11 +9,11 @@ public class StringCalculator {
 
     String[] split(String input) {
         validateDelimiters(input);
-        validateBlank(input);
         return input.split(" ");
     }
 
     private void validateDelimiters(String checkString){
+        checkString = checkString.trim();
         Pattern pattern = Pattern.compile(REGEX_PATTERN);
         if(!pattern.matcher(checkString).find()){
             throw new IllegalArgumentException("구분자는 공백을 사용합니다.");
@@ -32,6 +27,7 @@ public class StringCalculator {
     }
 
     public int calculate(String input) {
+        validateBlank(input);
         String[] stringArray = split(input);
         result = parseInt(stringArray[0]);
         for (int index = 1; index < stringArray.length; index++) {
@@ -53,43 +49,5 @@ public class StringCalculator {
 
     int calculate(int num, Operator operator, int num2) {
         return operator.calculator(num, num2);
-    }
-
-    enum Operator {
-        PLUS('+', (num1, num2) -> num1 + num2),
-        MINUS('-', (num1, num2) -> num1 - num2),
-        MULTIPLY('*', (num1, num2) -> num1 * num2),
-        DIVIDE('/', (num1, num2) -> {
-            validateDivision(num2);
-            return num1 / num2;
-        });
-
-        private final char symbol;
-        private final IntBinaryOperator expression;
-        private static final Map<String, Operator> symbolToEnum = Stream.of(values()).collect(Collectors.toMap(Operator::getSymbol, operator -> operator));
-
-        Operator(char symbol, IntBinaryOperator expression) {
-            this.symbol = symbol;
-            this.expression = expression;
-        }
-
-        public int calculator(int num1, int num2) {
-            return expression.applyAsInt(num1, num2);
-        }
-
-        private static void validateDivision(int number) {
-            if (number == 0) {
-                throw new IllegalArgumentException("0으로 나눌 수 없습니다.");
-            }
-        }
-
-        public String getSymbol() {
-            return String.valueOf(symbol);
-        }
-
-        public static Operator from(String symbol) {
-            return Optional.ofNullable(symbolToEnum.get(symbol))
-                    .orElseThrow(() -> new IllegalArgumentException(symbol + "은 지원하지 않는 연산기호입니다."));
-        }
     }
 }
